EA-9178: Remove trailing colon from library's RPATH
When we set OPENSSL_LIBRARIES and run cmake3, the file
./lib/CMakeFiles/zip.dir/link.txt is created.  It is in this
file the RPATH is set as -rpath,/opt/cpanel/ea-openssl11/lib:
I have yet to determine why the cmake3 infrastructure is
appending the colon to the OpenSSL libraries directory.
However, I can correct this between running cmake3 and make.
This fixes the RPATH in the library we build.